DRGameResult列挙型を別ファイルに移動。メッセージの設定処理をResultViewに任せるように修正。 #18
